requirements.txt
Pillow
numpy
tensorflow
streamlit